+++ jlox/src/main/java/com/ickoxii/jlox/Environment.java
@@ -1,4 +1,4 @@
-package com.ickoxii.loxinterpreter;
+package com.ickoxii.jlox;
 
 import java.util.HashMap;
 import java.util.Map;
