+++ jlox/src/main/java/com/ickoxii/jlox/LoxCallable.java
@@ -1,4 +1,4 @@
-package com.ickoxii.loxinterpreter;
+package com.ickoxii.jlox;
 
 import java.util.List;
